feat: mesh rendering system
Mesh and Shaders are now retrievable assets from a manager script.
With that, the mesh renderer creation can be handled within a class itself if it needs a mesh renderer.

This solves the issue of having to pass a mesh renderer as a parameter when spawning an Entity.

The meshRenderer variable is also now a pointer to save memory as certain entities require update but doesn't have visuals/graphics.