feat: boss cage class
It contains the basic implementation of a boss cage, which contains a wall grid and a floor grid. It contains the bounds that locks the player. Player locking within the cage isn't implemented yet.

Currently, it has the feature of highlighting player proximity on the grid so player can see walls clearer when they get closer to the wall.